Move MAGICCData back into pymagicc.io
